Structure for Simulator.run() method changed; check in initialize.
The way the Simulator.run() works has slightly changed to make the
execution safer plus to avoid concurrency problems with the stop()
method.

The initialize() method of the DEVSSimulator now checks for a running
simulation BEFORE it wipes the event list. One of the unit tests crashed
on the absence of this check.