End of replication now determined in Simulator run() method.
The END_REPLICATION_EVENT is fired by the SimulatorWorkerThread on the
basis of the ReplicationState.ENDING state. This used to be determined
by scheduling a simulation event on the eventlist at the end of the
runtime. This does, of course, not work for continuous simulations since
they do not have an eventlist. Also, the evnnt could be cancelled, or
not executed for other reasons. In the new setup, the run() method of
the Simulator determines whether the replication is over or not (it
tested for this situation anyhow). Therefore, the separate scheduling of
the end_repplication method is removed. end_replication can now be
CALLED from outside to end a replication before the official end.